+++ packages/fluent/modules/widgets/value/FluentUIDate.jsx
@@ -15,24 +15,33 @@ export default (props) => {
     placeholder,
   } = props;
 
+  const momentValue = value ? moment(value, valueFormat) : undefined;
+  const dateValue = momentValue ? momentValue.toDate() : undefined;
+
   const onChange = (date) => {
-    if (date == "" ) date = undefined;
-    var dateValue=moment(new Date(date)).format(valueFormat);
-    setValue(dateValue); 
+    // clear if invalid date
+    if (date == "" || date instanceof Date && isNaN(date))
+      date = undefined;
+    const newValue = date ? moment(date).format(valueFormat) : undefined;
+    setValue(newValue);
   };
+
+  const formatDate = (date) => moment(date).format(dateFormat);
 
   const stylesDatePicker = {
     // width: "auto", 
     marginRight: "0.25rem", 
     width: "150px"
   };
+
   return (
     <DatePicker
       disabled={readonly}
-      selectedDate={value}
+      value={dateValue}
       onSelectDate={onChange}
       style={stylesDatePicker}
       placeholder={placeholder}
+      formatDate={formatDate}
       {...customProps}
     />
   );